requirements.txt
Pillow==9.4.0

